fix: Split preview and expression from mapping-management.service (#2973)
Also moved getRuntimeVersion() to initialization service
Let mapping management service use file management service to upload mappings
First case for #2975, Moved `ITextNode` from `@atlasmap/atlasmap` to `@atlasmap/core` as `IExpressionNode`
