+++ ui/packages/atlasmap-core/src/services/document-management.service.ts
@@ -31,6 +31,7 @@ import { ConfigModel } from '../models/config.model';
 import { DocumentDefinition } from '../models/document-definition.model';
 import { DocumentInspectionModel } from '../models/inspect/document-inspection.model';
 import { DocumentInspectionUtil } from '../utils/document-inspection-util';
+import { Field } from '../models/field.model';
 import { Guid } from '../utils';
 import { HTTP_STATUS_NO_CONTENT } from '../common/constants';
 import ky from 'ky';
@@ -44,8 +45,7 @@ export class DocumentManagementService {
   cfg!: ConfigModel;
 
   private mappingUpdatedSubscription!: Subscription;
-
-  private headers = { 'Content-Type': 'application/json' };
+  private MAX_SEARCH_MATCH = 10000;
 
   constructor(private api: typeof ky) {}
 
@@ -428,5 +428,101 @@ export class DocumentManagementService {
         object: error,
       })
     );
+  }
+
+  /**
+   * Filter Document fields that is shwon in a Source/Target Document tree.
+   * @todo Consolidate with expression field search and Document Details field
+   * search - https://github.com/atlasmap/atlasmap/issues/603
+   * @param searchFilter
+   * @param isSource
+   */
+  filterDocumentFields(searchFilter: string | undefined, isSource: boolean) {
+    const cfg = ConfigModel.getConfig();
+
+    let searchResultsExist = false;
+    const searchIsEmpty: boolean =
+      undefined === searchFilter || '' === searchFilter;
+    const defaultVisibility: boolean = searchIsEmpty;
+    for (const docDef of cfg.getDocs(isSource)) {
+      docDef.visibleInCurrentDocumentSearch = defaultVisibility;
+      for (const field of docDef.getAllFields()) {
+        field.visibleInCurrentDocumentSearch = defaultVisibility;
+      }
+      if (!searchIsEmpty) {
+        let searchFieldCount = 0;
+        for (const field of docDef.getAllFields()) {
+          // Skip this field if it's already determined to be visible.
+          if (field.visibleInCurrentDocumentSearch && !field.collapsed) {
+            continue;
+          }
+          field.visibleInCurrentDocumentSearch = field.name
+            .toLowerCase()
+            .includes(searchFilter!.toLowerCase());
+          searchResultsExist =
+            searchResultsExist || field.visibleInCurrentDocumentSearch;
+
+          // The current field matches the user-specified filter.
+          if (field.visibleInCurrentDocumentSearch) {
+            docDef.visibleInCurrentDocumentSearch = true;
+            let parentField = field.parentField;
+
+            // Direct lineage is then visible.
+            while (
+              parentField != null &&
+              !parentField.visibleInCurrentDocumentSearch
+            ) {
+              parentField.visibleInCurrentDocumentSearch = true;
+              parentField.collapsed = false;
+              parentField = parentField.parentField;
+              searchFieldCount++;
+            }
+
+            // All fields below the matching field are also visible.
+            try {
+              this.markChildrenVisible(field);
+            } catch (error) {
+              cfg.errorService.addError(
+                new ErrorInfo({
+                  message: error.message,
+                  level: ErrorLevel.INFO,
+                  scope: ErrorScope.APPLICATION,
+                  type: ErrorType.USER,
+                })
+              );
+              break;
+            }
+
+            // The total number of matches is limited to allow the UI to perform.
+            if (searchFieldCount++ >= this.MAX_SEARCH_MATCH) {
+              cfg.errorService.addError(
+                new ErrorInfo({
+                  message:
+                    'The maximum number of fields matching the specified search filter has beeen exceeded  ' +
+                    'Try using a longer field filter.',
+                  level: ErrorLevel.INFO,
+                  scope: ErrorScope.APPLICATION,
+                  type: ErrorType.USER,
+                })
+              );
+              break;
+            }
+          }
+        }
+      }
+    }
+    cfg.mappingService.notifyLineRefresh();
+  }
+
+  private markChildrenVisible(field: Field): void {
+    field.visibleInCurrentDocumentSearch = true;
+    field.collapsed = false;
+    // if (this.searchFieldCount++ >= this.maxSearchMatch) {
+    //   throw new Error('The maximum number of fields matching the specified search filter has beeen exceeded  ' +
+    //     'Try using a longer field filter.');
+    // }
+    for (const childField of field.children) {
+      this.markChildrenVisible(childField);
+    }
   }
 }
