+++ ui/packages/atlasmap-core/src/utils/index.ts
@@ -14,8 +14,7 @@
     limitations under the License.
 */
 export * from './common-util';
-export * from './expression-util';
-export * from './filter-fields';
+export * from './document-inspection-util';
 export * from './guid';
 export * from './lookup-table-util';
 export * from './mapping-digest-util';
