+++ ui/packages/atlasmap/tsconfig.json
@@ -7,6 +7,7 @@
     "lib": ["dom", "esnext"],
     "importHelpers": true,
     "declaration": true,
+    "declarationMap": true,
     "sourceMap": true,
     "sourceRoot": "./src",
     "strict": true,
