+++ ui/test-resources/mapping/atlasmapping-expr-prop.json
@@ -163,7 +163,7 @@
               "@type" : "Capitalize"
             } ],
             "value" : "boston",
-            "docId" : "DOC.Properties.185063",
+            "docId" : "DOC.Properties",
             "index" : 0,
             "path" : "/prop-city",
             "fieldType" : "STRING",
@@ -172,7 +172,7 @@
           }, {
             "jsonType" : "io.atlasmap.v2.ConstantField",
             "value" : "Boston",
-            "docId" : "DOC.Constants.509124",
+            "docId" : "DOC.Constants",
             "index" : 1,
             "path" : "/Boston",
             "fieldType" : "STRING"
